% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{deterministic_cycles}
\alias{deterministic_cycles}
\title{Title}
\usage{
deterministic_cycles(arima_fit, time_instants = NULL, K = c(), period = c())
}
\arguments{
\item{arima_fit}{...}

\item{time_instants}{...}
}
\value{
...
}
\description{
Description ...
}
