% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_arma.R
\name{fit_arma}
\alias{fit_arma}
\title{Fit an ARMA model}
\usage{
fit_arma(ytdata)
}
\arguments{
\item{ytdata}{An object of class \code{\link{ytdata}}.}
}
\value{
An object of class \code{\link{ytfit}}.
}
\description{
This function ...
}
