% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.R
\name{ll_hmm}
\alias{ll_hmm}
\title{HMM log-likelihood}
\usage{
ll_hmm(theta, x, N, dist = "gaussian", neg = FALSE)
}
\arguments{
\item{theta}{A \code{numeric} parameter vector, see \code{\link{sample_theta}}.}

\item{x}{A \code{numeric} vector of observed data.}

\item{N}{An \code{integer}, the number of states in the HMM.}

\item{dist}{A \code{character} indicating the type of state-dependent distribution.
Can be \code{"gaussian"}, \code{"gamma"}, or \code{"poisson"}.}
}
\value{
The negative log-likelihood of the observed data.
}
\description{
Compute the log-likelihood of a Hidden Markov Model with state-dependent
Gaussian, Gamma, or Poisson distributions.
}
\examples{
theta <- c()
x <- c(1,2,NA,4,5)
ll_hmm(theta = theta, x = x, N = 2, dist = "poisson")

}
