% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.R
\name{mle_hmm}
\alias{mle_hmm}
\title{Maximum likelihood estimation of an HMM}
\usage{
mle_hmm(x, N, dist, runs = 1)
}
\arguments{
\item{x}{A \code{numeric} vector of observed data.}

\item{N}{An \code{integer}, the number of states. Must be greater or equal 2.}

\item{dist}{A \code{character} indicating the type of state-dependent distribution.
Can be \code{"gaussian"}, \code{"gamma"}, or \code{"poisson"}.}

\item{runs}{An \code{integer}, the number of randomly initialized optimization runs.}
}
\value{
The estimated parameter vector.
}
\description{
Numerical optimization of the Hidden Markov model likelihood function via
\code{\link[stats]{optim}}.
}
\examples{
\dontrun{
N <- 2
dist <- "gaussian"
theta <- sample_theta(N = N, dist = dist)
x <- simulate_hmm(T = 100, N = N, theta = theta, dist = dist)
mle <- mle_hmm(x = x, N = N, dist = dist, runs = 5)
}

}
