% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.ytdata.R
\name{plot.ytdata}
\alias{plot.ytdata}
\title{Plot method for a \code{ytdata} object.}
\usage{
\method{plot}{ytdata}(x, column, from = NA, to = NA, ...)
}
\arguments{
\item{x}{An object of class \code{\link{ytdata}}.}

\item{column}{A \code{character}, the column name of \code{x} to plot.}

\item{from}{A \code{character} of the form \code{"YYYY-MM-DD"}, the lower date.
Can be \code{NA} for no lower bound.}

\item{to}{A \code{character} of the form \code{"YYYY-MM-DD"}, the upper date.
Can be \code{NA} for no upper bound.}

\item{...}{Currently not used.}
}
\value{
A \code{ggplot} object.
}
\description{
This function is the plotting method for a \code{ytdata} object to visualize
the time series data.
}
