% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ytdata.R
\name{read_yt}
\alias{read_yt}
\title{Get data from YouTube Analytics API}
\usage{
read_yt(email, from = Sys.Date() - 10, to = Sys.Date())
}
\arguments{
\item{email}{A \code{character}, the email of the Google account.}

\item{from}{A \code{character} in \code{"YYYY-MM-DD"} format, the start date for
fetching YouTube Analytics data.
By default, \code{Sys.Date() - 10}.}

\item{to}{A \code{character} in \code{"YYYY-MM-DD"} format, the end date for
fetching YouTube Analytics data.
By default, \code{Sys.Date()}.}
}
\value{
An object of class \code{\link{ytdata}}.
}
\description{
Use this function to download YouTube Analytics API data.
}
