% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.R
\name{sample_theta}
\alias{sample_theta}
\title{Sample parameter vector for an HMM}
\usage{
sample_theta(N, dist, x = NULL)
}
\arguments{
\item{N}{An \code{integer}, the number of states. Must be greater or equal 2.}

\item{dist}{A \code{character} indicating the type of state-dependent distribution.
Can be \code{"gaussian"}, \code{"gamma"}, or \code{"poisson"}.}

\item{x}{A \code{numeric} vector of observed data.}
}
\value{
A \code{numeric} parameter vector.
The first \code{N*(N-1)} elements are the off-diagonal transition
probabilities.
The next \code{N} elements are the means.
Only if \code{dist = "gaussian"} or \code{dist = "gamma"}, the next \code{N}
elements are the standard deviations.
}
\description{
Sample the parameter vector \code{theta} for a Hidden Markov model.
}
\details{
If the observation vector \code{x} is supplied, means and standard deviations
are scaled by the first and second order data moments.
}
\examples{
\dontrun{
sample_theta(N = 3, dist = "gamma", x = 1:10)
}

}
