% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.R
\name{separate_theta}
\alias{separate_theta}
\title{Separate parameter vector}
\usage{
separate_theta(theta, N, dist, sort_states = TRUE)
}
\arguments{
\item{theta}{A \code{numeric} parameter vector, see \code{\link{sample_theta}}.}

\item{N}{An \code{integer}, the number of states. Must be greater or equal 2.}

\item{dist}{A \code{character} indicating the type of state-dependent distribution.
Can be \code{"gaussian"}, \code{"gamma"}, or \code{"poisson"}.}

\item{sort_states}{Set to \code{TRUE} to sort the states by their mean.}
}
\value{
A \code{list} of parameters.
}
\description{
Separate parameter vector \code{theta} into transition probability matrix
\code{Gamma}, stationary distribution \code{delta}, mean vector \code{mean},
and standard deviations \code{sd} (not if \code{dist = "poisson"}).
}
\examples{
\dontrun{
N <- 2
dist <- "poisson"
theta <- sample_theta(N = N, dist = dist)
separate_theta(theta = theta, N = N, dist = dist)
}
}
