% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_hmm.R
\name{sim_hmm}
\alias{sim_hmm}
\title{Simulate HMM}
\usage{
sim_hmm(T, N, Gamma, mu, sigma)
}
\description{
TODO
}
\examples{
\dontrun{
  Gamma <- matrix(c(0.9,0.1,0.1,0.9),nrow=2)
  mu <- c(1,3)
  sigma <- c(0.5,0.5)
  ts <- sim_hmm(T = 100, N = 2, Gamma = Gamma, mu = mu, sigma = sigma)
  model <- fit_hmm(ts, N = 2)
  states <- decode_states(ts, model, N = 2)
  plot(ts, type = "l")
  points(ts, col = states)
}
}
