% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.R
\name{simulate_hmm}
\alias{simulate_hmm}
\title{Simulate HMM data}
\usage{
simulate_hmm(T, N, theta, dist)
}
\arguments{
\item{T}{An \code{integer}, the time series length. Must be greater or equal 2.}

\item{N}{An \code{integer}, the number of states. Must be greater or equal 2.}

\item{theta}{A \code{numeric} parameter vector, see \code{\link{sample_theta}}.}

\item{dist}{A \code{character} indicating the type of state-dependent distribution.
Can be \code{"gaussian"}, \code{"gamma"}, or \code{"poisson"}.}
}
\value{
A \code{numeric} vector of length \code{T} with observations. The true state
sequence is available as attribute \code{"states"}.
}
\description{
Simulate time series data from a pre-specified Hidden Markov model.
}
\examples{
\dontrun{
N <- 2
dist <- "poisson"
theta <- sample_theta(N = N, dist = dist)
simulate_hmm(T = 20, N = N, theta = theta, dist = dist)
}

}
