% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{statistik_mit_lennart}
\alias{statistik_mit_lennart}
\title{Example YouTube data}
\format{
A \code{\link{ytdata}} object.
}
\usage{
statistik_mit_lennart
}
\description{
Data from the YouTube channel "Statistik mit Lennart"
(https://www.youtube.com/@statistik-mit-lennart) for demonstration.
}
\details{
Obtained via
\preformatted{
statisik_mit_lennart <- read_yt(
  email = "oelschlaeger.lennart@gmail.com",
  from = "2020-01-01",
  to = "2022-09-01"
)
}
}
\keyword{data}
