% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ytdata.R
\name{ytdata}
\alias{ytdata}
\alias{summary.ytdata}
\title{Constructor of a \code{ytdata} object}
\usage{
ytdata(data, channel, videos)

\method{summary}{ytdata}(object, ...)
}
\arguments{
\item{data}{A \code{tibble} with 6 columns and a row for each day and video:
\describe{
\item{day}{the date (\code{Date})}
\item{id}{the video id (\code{character})}
\item{views}{the number of views (\code{integer})}
\item{likes}{the number of likes (\code{integer})}
\item{dislikes}{the number of dislikes (\code{integer})}
\item{viewmins}{the number of minutes watched (\code{numeric})}
}}

\item{channel}{A \code{list} with 5 elements:
\describe{
\item{id}{the channel id (\code{character})}
\item{title}{the channel title (\code{character})}
\item{desc}{the channel description (\code{character})}
\item{publ}{the channel publication date (\code{Date})}
\item{views}{the total number of views (\code{integer})}
\item{subs}{the total number of subscribers (\code{integer})}
}}

\item{videos}{A \code{tibble} with 5 columns and a row for  each video:
\describe{
\item{id}{the video id (\code{character})}
\item{pub}{the date and time of publication (\code{POSIXct})}
\item{title}{the video title (\code{character})}
\item{desc}{the video description (\code{character})}
\item{tn}{the URL to the video thumbnail (\code{character})}
}}

\item{object}{An object of class \code{\link{ytdata}}.}

\item{...}{Currently not used.}
}
\value{
An object of class \code{\link{ytdata}}, which is the input \code{data}
with two attributes \code{channel} and \code{videos}.
}
\description{
This function creates a \code{ytdata} object which contains YouTube data
for modeling.
}
\examples{
print(statistik_mit_lennart)
summary(statistik_mit_lennart)
attributes(statistik_mit_lennart)$videos  # access video data
attributes(statistik_mit_lennart)$channel # access channel data


}
