% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ytdata_example}
\alias{ytdata_example}
\title{Example YouTube data}
\format{
A \code{\link{ytdata}} object.
}
\usage{
ytdata_example
}
\description{
Statistics from the YouTube channel "Statistik mit Lennart" for demonstration
}
\keyword{data}
