% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ytstats-package.R
\docType{package}
\name{ytstats-package}
\alias{ytstats}
\alias{ytstats-package}
\title{ytstats: Time Series Analysis of YouTube Data}
\description{
Implementation of different visualization and modeling tools for YouTube time series data like viewtimes or subscribers count.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/loelschlaeger/ytstats}
  \item Report bugs at \url{https://github.com/loelschlaeger/ytstats/issues}
}

}
\author{
\strong{Maintainer}: Lennart Oelschläger \email{oelschlaeger.lennart@gmail.com} (\href{https://orcid.org/0000-0001-5421-9313}{ORCID})

Authors:
\itemize{
  \item Rainer Buschmeier \email{rainer.buschmeier@uni-bielefeld.de}
}

}
\keyword{internal}
