% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.R
\name{decode_states}
\alias{decode_states}
\title{Decode HMM states}
\usage{
decode_states(x, theta, dist)
}
\arguments{
\item{x}{A \code{numeric} vector of observed data.}

\item{theta}{A \code{numeric} parameter vector, see \code{\link{sample_theta}}.}

\item{dist}{A \code{character} indicating the type of state-dependent distribution.
Can be \code{"gaussian"}, \code{"gamma"}, or \code{"poisson"}.}
}
\value{
TODO
}
\description{
TODO
}
\examples{
# TODO

}
