% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hmm.R
\name{sample_theta}
\alias{sample_theta}
\title{Sample parameter vector}
\usage{
sample_theta(N, dist, restricted = FALSE)
}
\arguments{
\item{N}{An \code{integer}, the number of states in the HMM.}

\item{dist}{A \code{character} indicating the type of state-dependent distribution.
Can be \code{"gaussian"}, \code{"gamma"}, or \code{"poisson"}.}
}
\value{
A \code{numeric} parameter vector.
The first \code{N*(N-1)} elements are the off-diagonal transition
probabilities.
The next \code{N} elements are the means.
Only if \code{dist = "gaussian"} or \code{dist = "gamma"}, the next \code{N}
elements are the standard deviations.
}
\description{
Sample the parameter vector \code{theta} for an HMM.
}
\examples{
sample_hmm_theta(N = 3, dist = "gamma")

}
